pyproject.toml
[project]
name = "sii-empresas-data-analysis"
version = "0.1.0"
description = "ETL pipeline for residuos peligrosos RNP 2023 with clean-code practices."
authors = [{ name = "Data Engineering Team" }]
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.2,<3",
    "pandera[pandas]>=0.26,<0.27",
    "pyarrow>=16,<17",
    "PyYAML>=6,<7",
    "plotly>=5.22,<6",
    "numpy>=1.26,<3",
    "python-dotenv>=1.0,<2",
]

[project.optional-dependencies]
dev = ["pytest>=8,<9"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra"
testpaths = ["tests"]
